requirements.txt
flask
flask-cors
google-genai
firebase-admin
gunicorn
